rework id to int for better hashcode

+++ de.tbonsack.karpfediem.pokemon.cardmanager.model/src/de/tbonsack/karpfediem/pokemon/cardmanager/model/objects/CardSet.java
@@ -10,7 +10,7 @@ public class CardSet extends ABasicCardObject implements Serializable {
 
 	private Set<Integer> _cards = new HashSet<>();
 
-	public CardSet(long id, String name, String imgPath) {
+	public CardSet(int id, String name, String imgPath) {
 		super(id, name, imgPath);
 	}
